backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
httpx==0.25.2
aiofiles==23.2.1
python-dotenv==1.0.0
Pillow==10.1.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
azure-storage-blob==12.19.0
azure-search-documents==11.4.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-monitor-opentelemetry==1.2.0
openai==1.6.1
langchain==0.0.352
langchain-openai==0.0.2
langchain-community==0.0.12
langgraph==0.0.25
tiktoken==0.5.2
azure-cosmos==4.5.1
azure-servicebus==7.11.4
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.1
isort==5.13.2
mypy==1.8.0
flake8==6.1.0
pre-commit==3.6.0
gunicorn==21.2.0
prometheus-client==0.19.0
structlog==23.2.0
